Use the default upgrader for feedback preparation if resume file not found
When the stage resume file is missing after completion, we are unable to detect the upgrader.
However, customers often encounter issues after the conversion is complete,
so we should be able to create a feedback archive even after the stage is finished.
Additionally, when there is only one upgrader, there seems to be no choice in which upgrader to use.