Added the `--local` option to reset the locale
This is useful in resolving encoding issues that may arise when the environment settings
are not compatible with utf-8 encoded files on the system.
For example there could be a problem when default values of the LANG or LC_CTYPE changed.

The option is preferable to changing environment variables as it ensures
that the locale will be adjusted within the script after reboot.

The option is not exposed to everyone due to the complexities and uncertainties
associated with resolving locale inconsistency problems. So we could advice to use
the option, but should not highlit this option in the help message.